requirements.txt
google-generativeai>=0.3.0
marshmallow>=3.0.0
PyJWT>=2.0
pytest==7.4.2
pytest-cov==4.1.0
flake8==6.0.0
Flask
Flask-Caching
Flask-Cors
Flask-SocketIO
Flask-SQLAlchemy
Flask-Migrate
Flask-Mail
Flask-Limiter>=3.0.0
celery[redis]>=5.0.0
redis>=4.0.0
python-dotenv
Flask-Babel
firebase-admin
requests
python-magic
bcrypt
email-validator
torch
tensorflow
keras
scikit-learn
joblib
xgboost
catboost
SpeechRecognition
deep-translator
streamlit
matplotlib
seaborn
opencv-python
pillow
reportlab
werkzeug
qrcode[pil]>=7.4.2
cryptography>=41.0.0
